Move grouped value handling to study-api
